Creating one link to the set resources
Instead of creating tow links to each resource ( the general site and the specific page for the topic), the suggestion can be made for just one link. Besides the user can look around if they want to, but they most likely click on the link to  get information about the specific topic ... not a general site.